requirements.txt
django
django-browser-reload
whitenoise
mysqlclient
db-sqlite3
django-widget-tweaks
livereload
django-debug-toolbar
pillow
django-cors-headers
django-livereload
django-simple-captcha
python-decouple
libsass
python-memcached
django-robots
django-htmlmin
django-cachalot
django-extensions
django-libsass
opencv-python
scikit-image
termcolor
django-sass-processor
django-compressor
stripe
djangorestframework
django-allauth
psycopg2-binary
django-cors-headers
slippers
celery
django-celery-beat
django-celery-results
redis